Create Cabins and CabinDetails page
The cabins page displays the cabins from the
cabins collection in the database. The cabin
details page sends a request with a route
parameter and shows the details of a specific
cabin. It also utilizes the Google Maps JavaScript
API to render a map on the page, highlighting the
cabin's location.