Create Authentication component
I created a LoginContext component which contains
the login and logout functions. The authentication
component renders a login or sign up form which
sends a request to the backend to receive the user
data and token as a response. The user data and
token will be saved as state in the LoginContext
component and in local storage.