Fix date filters and modals
I created an UnavailableBookingPeriodModal, which
is displayed when you change the date filters
inside CabinNavigation. The modal appears when a
cabinDetails page is already open but the cabin is
no longer available for the new dates. I also
added "Check Availability" and "Reset" buttons to
CabinNavigation to apply or reset the date
filters. This ensures that the date filters are
applied immediately after dates are selected and
helps fix potential bugs. Additionally, I
implemented functionality in the Modal component
to allow closing modals by clicking outside the
modal.